restructured
moved some function and classes to utils.
Tested and working .
