--- hyperparameters.yaml
@@ -1,3 +1,0 @@
-learning_rates: [0.001, 0.002, 0.003]
-latent_dims: [6, 8, 12]
-batch_sizes: [128, 256]
